FYAM-291: Your Data: shown UserData in page header. Updated network fetch logic.

+++ ForYouAndMe/Classes/Entities/UserData.swift
@@ -15,7 +15,7 @@ struct UserData {
     var title: String?
     @NilIfEmptyString
     var body: String?
-    let stars: Int
+    let stars: Double
 }
 
 extension UserData: JSONAPIMappable {
